+++ HISTORY.md
@@ -2,6 +2,18 @@
 
 ---
 
-## 0.0.0
+## 1.0.0
 
 `new` It is the first version of acharts.
+
+## 1.0.1
+
+1. custom tooltip
+2. pie’s placeholder
+
+## 1.0.2
+
+1. tooltip formatter
+2. custom labels
+3. plot events
+